First commit - Simple GUI
This is a simple gui that show the image in a fast manner.
They accept the S key for saving, the R key for refresh (should not
be need for this, but the gui is a stub, so should be buggy).
The executable is executed simply passing the scene name (do not used
yet) and the output image name. The file is saved automatically when the
rendering is finished.